requirements.txt
numpy
librosa
matplotlib
Pillow
soundfile
